Don't cut bones from within a hierarchy when deleting

+++ src/components/Anim.cpp
@@ -29,8 +29,11 @@ bool AnimInfo::RemoveShapeBone(const std::string& shape, const std::string& bone
 	shapeSkinning[shape].RemoveBone(boneName);
 
 	AnimSkeleton::getInstance().ReleaseBone(boneName);
-	if (AnimSkeleton::getInstance().GetBoneRefCount(boneName) <= 0)
-		refNif->DeleteNode(boneName);
+	if (AnimSkeleton::getInstance().GetBoneRefCount(boneName) <= 0) {
+		auto nodeParent = refNif->GetParentNode(refNif->FindBlockByName<NiNode>(boneName));
+		if (nodeParent == refNif->GetRootNode())
+			refNif->DeleteNode(boneName);
+	}
 
 	return true;
 }
@@ -40,8 +43,11 @@ void AnimInfo::Clear() {
 		for (auto &shapeBoneList : shapeBones) {
 			for (auto &boneName : shapeBoneList.second) {
 				AnimSkeleton::getInstance().ReleaseBone(boneName);
-				if (AnimSkeleton::getInstance().GetBoneRefCount(boneName) <= 0)
-					refNif->DeleteNode(boneName);
+				if (AnimSkeleton::getInstance().GetBoneRefCount(boneName) <= 0) {
+					auto nodeParent = refNif->GetParentNode(refNif->FindBlockByName<NiNode>(boneName));
+					if (nodeParent == refNif->GetRootNode())
+						refNif->DeleteNode(boneName);
+				}
 			}
 		}
 
@@ -59,8 +65,11 @@ void AnimInfo::ClearShape(const std::string& shape) {
 
 	for (auto &boneName : shapeBones[shape]) {
 		AnimSkeleton::getInstance().ReleaseBone(boneName);
-		if (AnimSkeleton::getInstance().GetBoneRefCount(boneName) <= 0)
-			refNif->DeleteNode(boneName);
+		if (AnimSkeleton::getInstance().GetBoneRefCount(boneName) <= 0) {
+			auto nodeParent = refNif->GetParentNode(refNif->FindBlockByName<NiNode>(boneName));
+			if (nodeParent == refNif->GetRootNode())
+				refNif->DeleteNode(boneName);
+		}
 	}
 
 	shapeBones.erase(shape);
